# Plugin/Swagger/bundled.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API OF CUDIDI",
    "description": "This is the documentation API of Cudidi browser",
    "contact": {
      "email": "Cudidi@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    },
    {
      "url": "https://localhost:3000/"
    }
  ],
...
  "paths": {
...
    "/article/create": {
      "post": {
        "tags": [
          "article"
        ],
        "summary": "Create article",
        "description": "Create article.",
        "operationId": "createArticle",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Create article",
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "Title": {
                    "type": "string"
                  },
                  "Location": {
                    "type": "string"
                  },
                  "Image": {
                    "type": "string"
                  },
                  "Duration": {
                    "type": "integer"
                  },
                  "Price": {
                    "type": "integer"
                  },
                  "NumberOfPeople": {
                    "type": "integer"
                  },
                  "Introduce": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "successful operation",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body"
      }
    },
    "/article/upload-image": {
      "post": {
        "tags": [
          "article"
        ],
        "summary": "Upload image of user",
        "description": "Upload image.",
        "operationId": "uploadImage",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Upload image",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Invalid user supplied",
            "content": {}
          },
          "401": {
            "description": "Access token is missing or invalid",
            "content": {}
          },
          "404": {
            "description": "User not found",
            "content": {}
          }
        }
      }
    },
    "/article/{id}": {
      "get": {
        "tags": [
          "article"
        ],
        "summary": "Get user by user id",
        "operationId": "getArticleById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id that needs to be fetched",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ID": {
                      "type": "integer",
                      "readOnly": true
                    },
                    "Username": {
                      "type": "string",
                      "minLength": 5,
                      "maxLength": 30
                    },
                    "Password": {
                      "type": "string",
                      "minLength": 6,
                      "writeOnly": true
                    },
                    "FullName": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 255
                    },
                    "Email": {
                      "type": "string",
                      "format": "email"
                    },
                    "BirthDay": {
                      "type": "string",
                      "format": "date"
                    },
                    "Address": {
                      "type": "string"
                    },
                    "Avatar": {
                      "type": "string"
                    },
                    "PhoneNumber": {
                      "type": "string",
                      "minLength": 5,
                      "maxLength": 15
                    }
                  },
                  "xml": {
                    "name": "User"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ID": {
                      "type": "integer",
                      "readOnly": true
                    },
                    "Username": {
                      "type": "string",
                      "minLength": 5,
                      "maxLength": 30
                    },
                    "Password": {
                      "type": "string",
                      "minLength": 6,
                      "writeOnly": true
                    },
                    "FullName": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 255
                    },
                    "Email": {
                      "type": "string",
                      "format": "email"
                    },
                    "BirthDay": {
                      "type": "string",
                      "format": "date"
                    },
                    "Address": {
                      "type": "string"
                    },
                    "Avatar": {
                      "type": "string"
                    },
                    "PhoneNumber": {
                      "type": "string",
                      "minLength": 5,
                      "maxLength": 15
                    }
                  },
                  "xml": {
                    "name": "User"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid article supplied",
            "content": {}
          },
          "404": {
            "description": "Article not found",
            "content": {}
          }
        }
      },
      "put": {
        "tags": [
          "article"
        ],
        "summary": "Updated article",
        "description": "Update information of article by id.",
        "operationId": "updateArticleById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "the id that need to be updated",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Updated article object",
          "content": {
            "*/*": {
              "schema": {
                "type": "object",
                "properties": {
                  "ID": {
                    "type": "integer",
                    "readOnly": true
                  },
                  "Location": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 255
                  },
                  "Image": {
                    "type": "string"
                  },
                  "Title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255
                  },
                  "Duration": {
                    "type": "integer"
                  },
                  "Price": {
                    "type": "number",
                    "minimum": 0,
                    "exclusiveMinimum": true
                  },
                  "NumberOfPeople": {
                    "type": "integer"
                  },
                  "Introduce": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 255
                  },
                  "descriptionArticles": {
                    "type": "object",
                    "properties": {
                      "ID": {
                        "type": "integer",
                        "readOnly": true
                      },
                      "Place": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 255
                      },
                      "Day": {
                        "type": "integer"
                      },
                      "Description": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 255
                      },
                      "imageArticles": {
                        "type": "object",
                        "properties": {
                          "ID": {
                            "type": "integer",
                            "readOnly": true
                          },
                          "URL": {
                            "type": "string",
                            "minLength": 1
                          },
                          "Location": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 255
                          }
                        },
                        "xml": {
                          "name": "Image-Article"
                        }
                      }
                    },
                    "xml": {
                      "name": "Description-Article"
                    }
                  }
                },
                "xml": {
                  "name": "Article"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid article supplied",
            "content": {}
          },
          "404": {
            "description": "Article not found",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body"
      },
      "delete": {
        "tags": [
          "article"
        ],
        "summary": "Delete article",
        "description": "Delete specify article by id.",
        "operationId": "deleleArticleById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id that needs to be deleted",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid article supplied",
            "content": {}
          },
          "404": {
            "description": "Article not found",
            "content": {}
          }
        }
      }
    },
    "/description-article/create": {
      "post": {
        "tags": [
          "description of article"
        ],
        "summary": "Create description of article",
        "description": "Create description day by day of article.",
        "operationId": "createDescriptionArticle",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Create description of article",
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "Article_Id": {
                    "type": "integer",
                    "readOnly": true
                  },
                  "Place": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255
                  },
                  "Day": {
                    "type": "integer"
                  },
                  "Description": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255
                  },
                  "imageArticles": {
                    "type": "object",
                    "properties": {
                      "ID": {
                        "type": "integer",
                        "readOnly": true
                      },
                      "URL": {
                        "type": "string",
                        "minLength": 1
                      },
                      "Location": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 255
                      }
                    },
                    "xml": {
                      "name": "Image-Article"
                    }
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "default": {
            "description": "successful operation",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body"
      }
    }
  }
}